Adds feature UploadAvatar
