+++ packages/client/src/features/UploadAvatar/index.ts
@@ -1,0 +1,1 @@
+export { UploadAvatar } from './ui';
