+++ packages/client/src/features/UploadAvatar/ui/UploadAvatar/UploadAvatar.constants.ts
@@ -1,0 +1,5 @@
+export const TEXTS = {
+  buttonText: 'Change Avatar',
+  avatarUpdateSuccess: 'Avatar updated successfully!',
+  avatarUpdateFailed: 'Failed to update avatar. Please try again.',
+} as const;
