+++ packages/client/src/widgets/ProfileAvatar/ui/ProfileAvatar/ProfileAvatar.interfaces.tsx
@@ -1,5 +1,3 @@
 export interface IProfileAvatarProps {
   avatarSrc: string;
-  handleBeforeUpload: (file: File) => boolean;
-  isLoading: boolean;
 }
